Fixing platecmodule to use uint32_t

+++ platec_src/platecmodule.cpp
@@ -4,6 +4,7 @@
 #endif
 #include <cmath>
 #include <Python.h>
+#include "utils.hpp"
 
 static PyObject * platec_create(PyObject *self, PyObject *args)
 {
@@ -57,9 +58,9 @@ PyObject *makelist(float array[], size_t size) {
     return l;
 }
 
-PyObject *makelist_int(size_t array[], size_t size) {
+PyObject *makelist_int(uint32_t array[], uint32_t size) {
     PyObject *l = PyList_New(size);
-    for (size_t i = 0; i != size; ++i) {
+    for (uint32_t i = 0; i != size; ++i) {
         PyList_SET_ITEM(l, i, Py_BuildValue("i",array[i]));
     }
     return l;
@@ -87,7 +88,7 @@ static PyObject * platec_get_platesmap(PyObject *self, PyObject *args)
     void *litho;
     if (!PyArg_ParseTuple(args, "l", &litho))
         return NULL; 
-    size_t *hm = platec_api_get_platesmap(litho);
+    uint32_t *hm = platec_api_get_platesmap(litho);
 
     size_t width = lithosphere_getMapWidth(litho);
     size_t height = lithosphere_getMapHeight(litho);
